requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
slack-sdk==3.31.0
google-genai==0.8.0
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
httpx==0.27.0
azure-storage-blob==12.22.0
python-multipart==0.0.9
